refactor: move state store to gitops/modelstore
