+++ internal/gitops/controller/network_reconciler_test.go
@@ -7,8 +7,8 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/swarm-deploy/swarm-deploy/internal/config"
-	"github.com/swarm-deploy/swarm-deploy/internal/gitops/controller/statem"
 	"github.com/swarm-deploy/swarm-deploy/internal/gitops/model"
+	"github.com/swarm-deploy/swarm-deploy/internal/gitops/modelstore"
 	"github.com/swarm-deploy/swarm-deploy/internal/labelsdict"
 	"github.com/swarm-deploy/swarm-deploy/internal/swarm"
 	"go.uber.org/mock/gomock"
@@ -144,7 +144,7 @@ func TestControllerSyncNetworksStoresState(t *testing.T) {
 		Create(gomock.Any(), gomock.Any()).
 		Return("created-id", nil)
 
-	store := statem.NewMemoryStore()
+	store := modelstore.NewMemoryStore()
 	c := &Controller{
 		cfg: &config.Config{
 			Spec: config.Spec{
@@ -185,7 +185,7 @@ func TestControllerSyncNetworksStoresFailedState(t *testing.T) {
 			Driver: "overlay",
 		}, nil)
 
-	store := statem.NewMemoryStore()
+	store := modelstore.NewMemoryStore()
 	c := &Controller{
 		cfg: &config.Config{
 			Spec: config.Spec{
@@ -214,7 +214,7 @@ func TestControllerSyncNetworksStoresFailedState(t *testing.T) {
 }
 
 func TestControllerSyncNetworksClearsStateWhenNetworksListIsEmpty(t *testing.T) {
-	store := statem.NewMemoryStore()
+	store := modelstore.NewMemoryStore()
 	store.Update(func(s *model.Runtime) {
 		s.Networks["legacy"] = model.Network{
 			Driver:     "overlay",
@@ -240,7 +240,7 @@ func TestControllerSyncNetworksClearsStateWhenNetworksListIsEmpty(t *testing.T) 
 }
 
 func TestControllerSyncNetworksSkipsReconcileWhenStateAlreadySyncedForCommit(t *testing.T) {
-	store := statem.NewMemoryStore()
+	store := modelstore.NewMemoryStore()
 	store.Update(func(s *model.Runtime) {
 		s.Networks["app_backend"] = model.Network{
 			Driver:     "overlay",
