+++ internal/gitops/modelstore/store_memory.go
@@ -1,4 +1,4 @@
-package statem
+package modelstore
 
 import (
 	"sync"
