+++ internal/gitops/modelstore/store_warmup.go
@@ -1,4 +1,4 @@
-package statem
+package modelstore
 
 import (
 	"log/slog"
